+++ cmd/env/env.go
@@ -6,8 +6,8 @@ import (
 	"github.com/spf13/cobra"
 
 	cmdpkg "github.com/rewardenv/reward/cmd"
+	"github.com/rewardenv/reward/internal/compose"
 	"github.com/rewardenv/reward/internal/config"
-	"github.com/rewardenv/reward/internal/dockercompose"
 	"github.com/rewardenv/reward/internal/logic"
 )
 
@@ -17,7 +17,7 @@ func NewCmdEnv(conf *config.Config) *cmdpkg.Command {
 			Use:                "env",
 			Short:              "Controls an environment from any point within the root project directory",
 			Long:               `Controls an environment from any point within the root project directory`,
-			ValidArgsFunction:  dockercompose.Completer(),
+			ValidArgsFunction:  compose.Completer(),
 			DisableFlagParsing: true,
 			RunE: func(cmd *cobra.Command, args []string) error {
 				err := logic.New(conf).RunCmdEnv(args)