+++ cmd/root/root.go
@@ -124,6 +124,9 @@ func NewCmdRoot(conf *config.Config) *cmdpkg.Command {
 }
 
 func configureFlags(cmd *cmdpkg.Command) {
+	cmd.ParseFlags(os.Args)
+	fmt.Println("args: ", os.Args)
+
 	// --app-dir
 	cmd.PersistentFlags().String(
 		"app-dir",
@@ -175,8 +178,8 @@ func configureFlags(cmd *cmdpkg.Command) {
 
 	// --driver
 	cmd.PersistentFlags().String(
-		"driver", "docker-compose", "orchestration driver")
-	_ = cmd.Config.BindPFlag(fmt.Sprintf("%s_driver", cmd.Config.AppName()), cmd.PersistentFlags().Lookup("driver"))
+		"driver", config.DriverDockerComposeV2, "orchestration driver")
+	cmd.Config.BindPFlag(fmt.Sprintf("%s_driver", cmd.Config.AppName()), cmd.PersistentFlags().Lookup("driver"))
 
 	// --service-domain
 	cmd.PersistentFlags().String(
@@ -264,7 +267,12 @@ func configureShortcuts(cmd *cmdpkg.Command) {
 
 func validateFlags(cmd *cmdpkg.Command) error {
 	driver := cmd.Config.GetString(fmt.Sprintf("%s_driver", cmd.Config.AppName()))
-	if !regexp.MustCompile(`^docker-compose$`).MatchString(driver) {
+	if !regexp.MustCompile(
+		fmt.Sprintf(`^%s|%s|%s$`,
+			config.DriverDockerCompose,
+			config.DriverDockerComposeV2,
+			config.DriverPodmanCompose,
+		)).MatchString(driver) {
 		return fmt.Errorf("invalid value for --driver: %s", driver)
 	}