+++ cmd/svc/svc.go
@@ -6,8 +6,8 @@ import (
 	"github.com/spf13/cobra"
 
 	cmdpkg "github.com/rewardenv/reward/cmd"
+	"github.com/rewardenv/reward/internal/compose"
 	"github.com/rewardenv/reward/internal/config"
-	"github.com/rewardenv/reward/internal/dockercompose"
 	"github.com/rewardenv/reward/internal/logic"
 )
 
@@ -17,7 +17,7 @@ func NewCmdSvc(conf *config.Config) *cmdpkg.Command {
 			Use:                "svc",
 			Short:              "Orchestrates global services such as traefik, portainer and dnsmasq via docker-compose",
 			Long:               `Orchestrates global services such as traefik, portainer and dnsmasq via docker-compose`,
-			ValidArgsFunction:  dockercompose.Completer(),
+			ValidArgsFunction:  compose.Completer(),
 			DisableFlagParsing: true,
 			RunE: func(cmd *cobra.Command, args []string) error {
 				err := logic.New(conf).RunCmdSvc(args)