+++ CHANGELOG.md
@@ -5,9 +5,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [0.5.3] - 2024-01-22
+## [0.6.0] - 2024-03-26
 
-We're super excited to announce `Reward` 0.5.3!
+We're super excited to announce `Reward` 0.6.0!
+
+### Changed
+
+- Add support to specify docker-compose version using the `--driver` flag
+- Default driver isA docker-compose v2
+
+## [0.5.3] - 2024-01-22
 
 ### Changed